Add better config inputting
