+++ gen/tokenize_data.py
@@ -11,7 +11,8 @@ import pickle
 import numpy as np
 import itertools
 from collections import Counter
-from config import rand_path, train_path, val_path, meta_path
+
+from config import rand_path, train_path, val_path, meta_path, TokenizeConfig, parse_args
 
 max_vocab_size = 64
 
@@ -31,13 +32,7 @@ def bpe_replace(L, a, b, c):
            itertools.chain.from_iterable(L[i+2:j] + [c] for i, j in zip(replace_idx[:-1], replace_idx[1:]))) + \
            L[replace_idx[-1]+2:]
 
-
-def main():
-    data = ""
-    with open(rand_path, 'r') as f:
-        data = f.read()
-    print(f"Total number of characters in dataset: {len(data):,}")
-
+def tokenize(data, args: TokenizeConfig):
     chars = sorted(list(set(data)))
     vocab_size = len(chars)
 
@@ -47,34 +42,59 @@ def main():
     stoi = { ch: i  for i, ch in enumerate(chars) }
     itos = { i : ch for i, ch in enumerate(chars) }
 
-    enc = {}
-    # list where bpe[i] = s iff s encodes to i
-    dec = list(chars)
 
-    encoded_data = [stoi[c] for c in data]
-    while vocab_size < max_vocab_size:
-        if len(encoded_data) < 2: break
+    if args.encoding == 'bpe':
+        enc = {}
+        # list where bpe[i] = s iff s encodes to i
+        dec = list(chars)
 
-        # slow but that's ok
-        counter = Counter(zip(encoded_data[:-1], encoded_data[1:]))
-        pair, _ = counter.most_common(1)[0]
+        encoded_data = [stoi[c] for c in data]
+        while vocab_size < max_vocab_size:
+            if len(encoded_data) < 2: break
 
-        enc[vocab_size] = pair
-        dec.append(dec[pair[0]] + dec[pair[1]])
-        print(f"Compressed pair: {(dec[pair[0]] + dec[pair[1]])!r}")
-        
-        encoded_data = bpe_replace(encoded_data, pair[0], pair[1], vocab_size)
-        vocab_size += 1
+            # slow but that's ok
+            counter = Counter(zip(encoded_data[:-1], encoded_data[1:]))
+            pair, _ = counter.most_common(1)[0]
 
-    # Encode and decode
-    def encode(s):
-        L = [stoi[c] for c in s]
-        for k, pair in enc.items():
-            L = bpe_replace(L, pair[0], pair[1], k)
-        return L
-        
-    #def decode(l):
-    #    return ''.join(dec[i] for i in l)
+            enc[vocab_size] = pair
+            dec.append(dec[pair[0]] + dec[pair[1]])
+            print(f"Compressed pair: {(dec[pair[0]] + dec[pair[1]])!r}")
+            
+            encoded_data = bpe_replace(encoded_data, pair[0], pair[1], vocab_size)
+            vocab_size += 1
+
+        # Encode and decode
+        def encode(s):
+            L = [stoi[c] for c in s]
+            for k, pair in enc.items():
+                L = bpe_replace(L, pair[0], pair[1], k)
+            return L
+            
+        #def decode(l):
+        #    return ''.join(dec[i] for i in l)
+
+        # Save meta information to help encoding/decoding later
+        meta = {
+            'args': args.__dict__,
+            'vocab_size': vocab_size,
+            'itos': itos,
+            'stoi': stoi,
+            'enc': enc,
+            'dec': dec,
+        }
+    elif args.encoding == 'char': 
+        def encode(s):
+            return [stoi[c] for c in s]
+        #def decode(l):
+        #    return ''.join(itos[i] for i in l)
+        meta = {
+            'args': args.__dict__,
+            'vocab_size': vocab_size,
+            'itos': itos,
+            'stoi': stoi,
+        }
+    else:
+        raise ValueError(f"Unknown encoding: {args.encoding}")
 
     #print(encoded_data == encode(data))
     #print(data == decode(encoded_data))
@@ -87,6 +107,20 @@ def main():
     # Encode to integers
     train_ids = encode(train_data)
     val_ids = encode(val_data)
+
+    return train_ids, val_ids, meta
+
+def main():
+    args = TokenizeConfig(**parse_args())
+    
+    data = ""
+    with open(rand_path, 'r') as f:
+        data = f.read()
+    print(f"Total number of characters in dataset: {len(data):,}")
+
+    train_ids, val_ids, meta = tokenize(data, args)
+    vocab_size = meta['vocab_size']
+
     print(f"Final vocab size: {vocab_size:,}")
     print(f"train has {len(train_ids):,} tokens")
     print(f"val has {len(val_ids):,} tokens")
@@ -96,15 +130,6 @@ def main():
     val_ids = np.array(val_ids, dtype=np.uint16)
     train_ids.tofile(train_path)
     val_ids.tofile(val_path)
-
-    # Save meta information to help encoding/decoding later
-    meta = {
-        'vocab_size': vocab_size,
-        'itos': itos,
-        'stoi': stoi,
-        'enc': enc,
-        'dec': dec,
-    }
 
     with open(meta_path, 'wb') as f:
         pickle.dump(meta, f)
